requirements.txt
flask
gunicorn
pillow
matplotlib
opencv-python-headless
numpy